pyproject.toml
[project]
name = "finance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-auth>=2.49.2",
    "groq>=1.1.2",
    "gspread>=6.2.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
